library/pyproject.toml
[project]
name = "library"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==4.2",
    "djangorestframework>=3.16.0",
]
